+++ src/main/java/de/deruser/kickertracker/service/algorithms/glicko/CompositeOpponent.java
@@ -70,10 +70,10 @@ public class CompositeOpponent implements StatsAlgorithm {
       Glicko2 newGlicko = computeGlicko(getStats(currentPlayerInfo, matchTimestamp), opponent, eloResult);
 
       Player.PlayerBuilder playerBuilder = player.toBuilder()
-          .glicko((int) newGlicko.rating())
-          .deviation((int) newGlicko.ratingDeviation())
+          .glicko(newGlicko.rating())
+          .deviation(newGlicko.ratingDeviation())
           .volatility(newGlicko.ratingVolatility())
-          .glickoChange((int) newGlicko.rating() - currentPlayerInfo.getGameStats().getGlicko());
+          .glickoChange(newGlicko.rating() - currentPlayerInfo.getGameStats().getGlicko());
       result.add(playerBuilder.build());
     }
     return result;
