+++ src/main/java/de/deruser/kickertracker/service/algorithms/glicko/Individual.java
@@ -70,10 +70,10 @@ public class Individual implements StatsAlgorithm {
               getStats(opponentPlayerInfo, matchTimestamp), eloResult);
 
       Player.PlayerBuilder playerBuilder = player.toBuilder()
-          .glicko((int) newGlicko.rating())
-          .deviation((int) newGlicko.ratingDeviation())
+          .glicko( newGlicko.rating())
+          .deviation( newGlicko.ratingDeviation())
           .volatility(newGlicko.ratingVolatility())
-          .glickoChange((int) newGlicko.rating() - currentPlayerInfo.getGameStats().getGlicko());
+          .glickoChange( newGlicko.rating() - currentPlayerInfo.getGameStats().getGlicko());
       result.add(playerBuilder.build());
     }
     return result;
@@ -109,8 +109,8 @@ public class Individual implements StatsAlgorithm {
         opponent = currentOpponent;
         continue;
       }
-      int currentOpponentGlicko = Math.abs(player.getGlicko() - currentOpponent.getGlicko());
-      int oldOpponentGlicko = Math.abs(player.getGlicko() - currentOpponent.getGlicko());
+      double currentOpponentGlicko = Math.abs(player.getGlicko() - currentOpponent.getGlicko());
+      double oldOpponentGlicko = Math.abs(player.getGlicko() - currentOpponent.getGlicko());
       if(currentOpponentGlicko < oldOpponentGlicko){
         opponent = currentOpponent;
       }
